Fix 401 error when publishing a new weblog entry
The add entry button in the toolbar wasn’t passing the selected address
to the lower layers. As result, Triton was making a network request to
create a new post without the address. In such cases, the API returns an
error, 401.

This commit ensures the address is getting passed down to the lower
layers before making the network request.